adding vision in its hope. added instructions
