+++ src/main/java/frc/robot/BuildConstants.java
@@ -7,13 +7,13 @@ public final class BuildConstants {
   public static final String MAVEN_GROUP = "";
   public static final String MAVEN_NAME = "Robot2025";
   public static final String VERSION = "unspecified";
-  public static final int GIT_REVISION = 55;
-  public static final String GIT_SHA = "bae3893e40511eb9f524cfefe7ca39610c95feac";
-  public static final String GIT_DATE = "2025-01-14 19:20:21 EST";
-  public static final String GIT_BRANCH = "main";
-  public static final String BUILD_DATE = "2025-01-17 17:12:57 EST";
-  public static final long BUILD_UNIX_TIME = 1737151977175L;
-  public static final int DIRTY = 0;
+  public static final int GIT_REVISION = 62;
+  public static final String GIT_SHA = "80c431edd04cf59ce250d8d2d3862522c3a16136";
+  public static final String GIT_DATE = "2025-01-18 09:44:04 EST";
+  public static final String GIT_BRANCH = "dc_vision_tweaks";
+  public static final String BUILD_DATE = "2025-01-18 13:33:25 EST";
+  public static final long BUILD_UNIX_TIME = 1737225205801L;
+  public static final int DIRTY = 1;
 
   private BuildConstants(){}
 }
