+++ vision/README.md
@@ -1,0 +1,20 @@
+# 2025 Vision Project
+
+## developing
+
+1. install python 3.11.x (must be this version!)
+
+Then:
+
+```cmdline
+pip install -r requirements.txt
+pytest
+```
+
+
+## Deploying to the robot
+```.\gradlew deployVision```
+
+
+## Building the PI Image
+See https://github.com/entech281/2025_opi_image
