+++ vision/install.bat
@@ -1,0 +1,1 @@
+pip install --extra-index-url=https://wpilib.jfrog.io/artifactory/api/pypi/wpilib-python-release-2025/simple -r requirements.txt
